корректирует файловую архитектуру, обновляет разметку reviews
