+++ src/components/tabs/details/details.jsx
