+++ src/components/tabs/details/details.test.js
